requirements.txt
# Core ML
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0

# Computer Vision / Feature Extraction
opencv-python>=4.8
scikit-image>=0.21

# Dataset
SoccerNet>=0.1.60

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Utilities
tqdm
Pillow